feat(svmgov): zero copy handling of supporters
