+++ svmgov/program/programs/svmgov_program/src/constants.rs
@@ -8,7 +8,8 @@ pub const MAX_TITLE_ACCOUNT_SIZE: usize = 200;
 pub const MAX_DESC_ACCOUNT_SIZE: usize = 500;
 
 // Hard upper bound on the admin-configurable `max_supporters` cap. Every support
-// call must deserialize and re-tally the whole `supporters` list in a single
-// transaction; 2000 staked validators is now the maximum after SIMD-0357:
-// Alpenglow VAT implementation has been activated on all clusters.
+// call must re-tally the whole supporter list (read zero-copy from the proposal
+// account) in a single transaction; 2000 staked validators is now the maximum
+// after SIMD-0357: Alpenglow VAT implementation has been activated on all
+// clusters.
 pub const MAX_SUPPORTERS_LIMIT: u32 = 2_000;
